feat(cli): remove `smithery skill` subcommand (SMI-1682)
`npx skills` (vercel-labs/skills) is the canonical skill installer, and our
`skill add` already just shelled out to it. Install-count attribution is
tracked server-side when the skill is fetched from smithery.ai — not in this
CLI — so removing the command preserves tracking (verified). This keeps the
CLI focused on tools/MCP.

- Remove the `skill` command group (agents/search/view/add/publish) + the
  `skills` plural alias from src/index.ts
- Delete src/commands/skill/ (search/view/publish/publish-utils/shared)
- Relocate the still-needed `installSkill` passthrough to
  src/lib/skill-install.ts (self-contained) — `setup` depends on it to
  install the Smithery CLI's own skill; repoint its import
- Move/trim tests to src/lib/__tests__/skill-install.test.ts (keeps
  installSkill + setup-flow coverage; drops deleted-command tests)
- README: point skill install at `npx skills add`

Pre-existing mcp/* typecheck errors and the uplink `ws` test failures are
baseline (identical on origin/main), unrelated to this change.
