--- src/commands/skill/index.ts
@@ -1,4 +1,0 @@
-export { installSkill } from "./install"
-export { publishSkill } from "./publish"
-export { searchSkills } from "./search"
-export { viewSkill } from "./view"
