--- src/commands/skill/search.ts
@@ -1,323 +1,0 @@
-import type { Smithery } from "@smithery/api"
-import type { SkillListResponse } from "@smithery/api/resources/skills"
-import pc from "picocolors"
-import { SKILL_AGENTS } from "../../config/agents.js"
-import { fatal } from "../../lib/cli-error"
-import { isJsonMode, outputTable, truncate } from "../../utils/output"
-import { installSkill } from "./install.js"
-import { createPublicSkillsClient, getSkillUrl } from "./shared.js"
-
-export interface SearchOptions {
-	interactive?: boolean
-	limit?: number
-	page?: number
-	namespace?: string
-}
-
-/**
- * Prompts for a search term if not provided
- */
-async function getSearchTerm(providedTerm?: string): Promise<string> {
-	if (providedTerm) {
-		return providedTerm
-	}
-
-	const inquirer = (await import("inquirer")).default
-	const result = await inquirer.prompt([
-		{
-			type: "input",
-			name: "searchTerm",
-			message: "Search for skills:",
-			validate: (input: string) =>
-				input.trim().length > 0 || "Please enter a search term",
-		},
-	])
-	console.log()
-	return result.searchTerm
-}
-
-/**
- * Format skill for display in autocomplete list
- */
-function formatSkillDisplay(skill: SkillListResponse): string {
-	const displayName = skill.displayName || `${skill.namespace}/${skill.slug}`
-	const stats =
-		skill.externalStars && skill.externalStars > 0
-			? `★ ${skill.externalStars.toLocaleString()}`
-			: ""
-	const qualifiedName = `${skill.namespace}/${skill.slug}`
-	const statsDisplay = stats ? ` • ${stats}` : ""
-
-	return `${displayName}${statsDisplay} • ${pc.dim(qualifiedName)}`
-}
-
-/**
- * Interactive skill search and browsing
- * @param initialQuery - Initial search query (optional)
- * @param options - Search options (dump mode, limit)
- * @returns Promise<SkillListResponse | null> - Selected skill or null if cancelled
- */
-export async function searchSkills(
-	initialQuery?: string,
-	options: SearchOptions = {},
-): Promise<SkillListResponse | null> {
-	const { interactive = false, limit = 10, page = 1, namespace } = options
-	const json = isJsonMode()
-
-	const searchTerm = interactive
-		? await getSearchTerm(initialQuery)
-		: (initialQuery ?? "")
-
-	try {
-		const client = createPublicSkillsClient()
-
-		// Build query params
-		const queryParams: {
-			q?: string
-			pageSize: number
-			page?: number
-			namespace?: string
-		} = {
-			pageSize: limit,
-			page,
-		}
-		if (searchTerm) {
-			queryParams.q = searchTerm
-		}
-		if (namespace) {
-			queryParams.namespace = namespace
-		}
-
-		if (interactive) {
-			return interactiveSearch(client, queryParams, limit, searchTerm)
-		}
-
-		const response = await client.skills.list(queryParams)
-		const skills = response.skills
-
-		const data = skills.map((skill) => ({
-			name: skill.displayName || `${skill.namespace}/${skill.slug}`,
-			qualifiedName: `${skill.namespace}/${skill.slug}`,
-			description: skill.description ?? "",
-			stars: skill.externalStars ?? 0,
-		}))
-
-		const jsonData = json
-			? {
-					skills: skills.map((skill) => {
-						const {
-							vector,
-							$dist,
-							score,
-							gitUrl,
-							totalActivations,
-							uniqueUsers,
-							externalStars,
-							...rest
-						} = skill as SkillListResponse & {
-							vector?: unknown
-							$dist?: unknown
-							score?: unknown
-						}
-						return {
-							...rest,
-							stars: externalStars,
-							url: getSkillUrl(skill.namespace, skill.slug),
-						}
-					}),
-				}
-			: undefined
-
-		const hasMore = skills.length >= limit
-
-		outputTable({
-			data,
-			columns: [
-				{ key: "name", header: "NAME" },
-				{ key: "qualifiedName", header: "QUALIFIED_NAME" },
-				{
-					key: "stars",
-					header: "STARS",
-					format: (v) => (Number(v) > 0 ? String(v) : ""),
-				},
-				{
-					key: "description",
-					header: "DESCRIPTION",
-					format: (v) => truncate(String(v ?? "")),
-				},
-			],
-			json,
-			jsonData,
-			pagination: { page, hasMore },
-			tip:
-				data.length === 0
-					? "No skills found. Try a different search term."
-					: "Use smithery skills install <qualified-name> --agent <agent> to install.",
-		})
-		return null
-	} catch (error) {
-		fatal("Error searching skills", error)
-	}
-}
-
-async function interactiveSearch(
-	client: Smithery,
-	queryParams: {
-		q?: string
-		pageSize: number
-		page?: number
-		namespace?: string
-	},
-	limit: number,
-	initialSearchTerm: string,
-): Promise<SkillListResponse | null> {
-	let searchTerm = initialSearchTerm
-
-	while (true) {
-		if (searchTerm) {
-			queryParams.q = searchTerm
-		}
-
-		const yoctoSpinner = (await import("yocto-spinner")).default
-		const searchMsg = queryParams.namespace
-			? `Searching in ${queryParams.namespace}${searchTerm ? ` for "${searchTerm}"` : ""}...`
-			: `Searching for "${searchTerm}"...`
-		const spinner = yoctoSpinner({ text: searchMsg }).start()
-
-		const response = await client.skills.list(queryParams)
-		const skills = response.skills
-
-		if (skills.length === 0) {
-			spinner.error(`No skills found for "${searchTerm}"`)
-			return null
-		}
-
-		spinner.success(
-			`☀ ${skills.length < limit ? `Found ${skills.length} result${skills.length === 1 ? "" : "s"}:` : `Showing top ${skills.length} results:`}`,
-		)
-		console.log(
-			pc.dim(
-				`${pc.cyan("→ View more")} at smithery.ai/skills?q=${searchTerm.replace(/\s+/g, "+")}`,
-			),
-		)
-		console.log()
-
-		const inquirer = (await import("inquirer")).default
-		const autocompletePrompt = (await import("inquirer-autocomplete-prompt"))
-			.default
-		inquirer.registerPrompt("autocomplete", autocompletePrompt)
-
-		const { selectedSkill } = await inquirer.prompt([
-			{
-				type: "autocomplete",
-				name: "selectedSkill",
-				message: "Select skill for details (or search again):",
-				source: (_: unknown, input: string) => {
-					const options = [
-						{ name: pc.dim("← Search again"), value: "__SEARCH_AGAIN__" },
-						{ name: pc.dim("Exit"), value: "__EXIT__" },
-					]
-
-					const filtered = skills
-						.filter((s) => {
-							const searchStr = (input || "").toLowerCase()
-							return (
-								s.displayName?.toLowerCase().includes(searchStr) ||
-								s.slug.toLowerCase().includes(searchStr) ||
-								s.namespace.toLowerCase().includes(searchStr) ||
-								s.description?.toLowerCase().includes(searchStr)
-							)
-						})
-						.map((s) => ({
-							name: formatSkillDisplay(s),
-							value: s.id,
-						}))
-
-					return Promise.resolve([...options, ...filtered])
-				},
-			},
-		])
-
-		if (selectedSkill === "__EXIT__") {
-			return null
-		} else if (selectedSkill === "__SEARCH_AGAIN__") {
-			searchTerm = await getSearchTerm()
-			continue
-		}
-
-		console.log()
-		const selectedSkillData = skills.find((s) => s.id === selectedSkill)
-		if (selectedSkillData) {
-			const displayName =
-				selectedSkillData.displayName ||
-				`${selectedSkillData.namespace}/${selectedSkillData.slug}`
-			console.log(`${pc.bold(pc.cyan(displayName))}`)
-			console.log(
-				`${pc.dim("Qualified name:")} ${selectedSkillData.namespace}/${selectedSkillData.slug}`,
-			)
-			if (selectedSkillData.description) {
-				console.log(
-					`${pc.dim("Description:")} ${selectedSkillData.description}`,
-				)
-			}
-			if (
-				selectedSkillData.categories &&
-				selectedSkillData.categories.length > 0
-			) {
-				console.log(
-					`${pc.dim("Categories:")} ${selectedSkillData.categories.join(", ")}`,
-				)
-			}
-			if (selectedSkillData.externalStars) {
-				console.log(
-					`${pc.dim("Stars:")} ${selectedSkillData.externalStars.toLocaleString()}`,
-				)
-			}
-			console.log()
-
-			const skillIdentifier = `${selectedSkillData.namespace}/${selectedSkillData.slug}`
-			console.log(pc.bold("To install this skill, run:"))
-			console.log()
-			console.log(
-				pc.cyan(
-					`  smithery skills install ${skillIdentifier} --agent <agent-name>`,
-				),
-			)
-			console.log()
-
-			const { action } = await inquirer.prompt([
-				{
-					type: "list",
-					name: "action",
-					message: "What would you like to do?",
-					choices: [
-						{ name: "↓ Install", value: "install" },
-						{ name: "← Back to skill list", value: "back" },
-						{ name: "← Search again", value: "search" },
-						{ name: "Exit", value: "exit" },
-					],
-				},
-			])
-
-			if (action === "install") {
-				const { agent } = await inquirer.prompt([
-					{
-						type: "list",
-						name: "agent",
-						message: "Select the agent to install to:",
-						choices: SKILL_AGENTS.map((a) => ({ name: a, value: a })),
-					},
-				])
-
-				await installSkill(skillIdentifier, agent)
-				return null
-			} else if (action === "back") {
-				console.log()
-			} else if (action === "search") {
-				searchTerm = await getSearchTerm()
-			} else {
-				return null
-			}
-		}
-	}
-}
