+++ src/lib/__tests__/skill-install.test.ts
@@ -1,6 +1,9 @@
 /**
- * Skills Command Tests
- * Validates that skills search and install use public API (no auth required)
+ * skill-install tests
+ *
+ * The user-facing `smithery skill` command was removed (SMI-1682); the
+ * `installSkill` passthrough is retained for the `setup` command. These tests
+ * lock its public-API resolve (no auth) and npx-skills flag mapping.
  */
 
 import { beforeEach, describe, expect, test, vi } from "vitest"
@@ -10,17 +13,6 @@ function createSmitheryMock() {
 		Smithery: vi.fn().mockImplementation((config) => ({
 			_config: config,
 			skills: {
-				list: vi.fn().mockResolvedValue({
-					skills: [
-						{
-							id: "test-id",
-							namespace: "test-ns",
-							slug: "test-skill",
-							displayName: "Test Skill",
-							description: "A test skill",
-						},
-					],
-				}),
 				get: vi.fn().mockResolvedValue({
 					id: "test-id",
 					namespace: "test-ns",
@@ -45,32 +37,20 @@ vi.spyOn(console, "log").mockImplementation(() => {})
 vi.spyOn(console, "error").mockImplementation(() => {})
 
 import { Smithery } from "@smithery/api"
-import { setOutputMode } from "../../utils/output"
 
 async function getCommand(): Promise<string> {
 	const { execSync } = await import("node:child_process")
 	return (execSync as ReturnType<typeof vi.fn>).mock.calls[0][0]
 }
 
-describe("skills commands use public API", () => {
+describe("installSkill resolves via public API (no auth)", () => {
 	beforeEach(() => {
 		vi.clearAllMocks()
 	})
 
-	test("skills search creates Smithery client with empty API key", async () => {
-		const { searchSkills } = await import("../skill/search")
-
-		// Run search in JSON mode to avoid interactive prompts
-		setOutputMode({ json: true })
-		await searchSkills("test", { limit: 5 })
-
-		// Verify Smithery was instantiated with empty API key
-		expect(Smithery).toHaveBeenCalledWith({ apiKey: "" })
-	})
-
-	test("skills install resolves skill URL with empty API key", async () => {
+	test("creates Smithery client with empty API key and delegates to npx skills", async () => {
 		const { execSync } = await import("node:child_process")
-		const { installSkill } = await import("../skill/install")
+		const { installSkill } = await import("../skill-install")
 
 		await installSkill("test-ns/test-skill", "claude-code", {})
 
@@ -88,7 +68,7 @@ describe("installSkill flag mapping", () => {
 	})
 
 	test("no agent, no options — interactive, no flags", async () => {
-		const { installSkill } = await import("../skill/install")
+		const { installSkill } = await import("../skill-install")
 		await installSkill("test-ns/test-skill")
 
 		const command = await getCommand()
@@ -98,7 +78,7 @@ describe("installSkill flag mapping", () => {
 	})
 
 	test("with agent, no yes — adds --agent but stays interactive", async () => {
-		const { installSkill } = await import("../skill/install")
+		const { installSkill } = await import("../skill-install")
 		await installSkill("test-ns/test-skill", "claude-code", {})
 
 		const command = await getCommand()
@@ -107,7 +87,7 @@ describe("installSkill flag mapping", () => {
 	})
 
 	test("with agent and yes — adds --agent and -y", async () => {
-		const { installSkill } = await import("../skill/install")
+		const { installSkill } = await import("../skill-install")
 		await installSkill("test-ns/test-skill", "claude-code", { yes: true })
 
 		const command = await getCommand()
@@ -116,7 +96,7 @@ describe("installSkill flag mapping", () => {
 	})
 
 	test("all options — maps every flag", async () => {
-		const { installSkill } = await import("../skill/install")
+		const { installSkill } = await import("../skill-install")
 		await installSkill("test-ns/test-skill", "cursor", {
 			global: true,
 			yes: true,
@@ -131,15 +111,16 @@ describe("installSkill flag mapping", () => {
 	})
 })
 
-describe("setup command flow", () => {
+describe("setup command flow (the remaining installSkill consumer)", () => {
 	beforeEach(() => {
 		vi.clearAllMocks()
 	})
 
 	test("setup passes global and yes by default", async () => {
-		const { installSkill } = await import("../skill/install")
+		const { installSkill } = await import("../skill-install")
 
-		// Mirrors what setup does: installSkill("smithery-ai/cli", undefined, { global: true, yes: true })
+		// Mirrors what `setup` does:
+		// installSkill("smithery-ai/cli", agent, { global: true, yes: true })
 		await installSkill("smithery-ai/cli", undefined, {
 			global: true,
 			yes: true,
@@ -148,33 +129,5 @@ describe("setup command flow", () => {
 		const command = await getCommand()
 		expect(command).toContain("-g")
 		expect(command).toMatch(/-y$/)
-	})
-})
-
-describe("skill install command flow", () => {
-	beforeEach(() => {
-		vi.clearAllMocks()
-	})
-
-	test("with agent — non-interactive (yes: true)", async () => {
-		const { installSkill } = await import("../skill/install")
-
-		// Mirrors: installSkill(skill, agent, { global, yes: !!agent })
-		await installSkill("test-ns/test-skill", "claude-code", { yes: true })
-
-		const command = await getCommand()
-		expect(command).toContain("--agent claude-code")
-		expect(command).toMatch(/-y$/)
-	})
-
-	test("without agent — interactive (yes: false)", async () => {
-		const { installSkill } = await import("../skill/install")
-
-		// Mirrors: installSkill(skill, undefined, { global, yes: false })
-		await installSkill("test-ns/test-skill", undefined, { yes: false })
-
-		const command = await getCommand()
-		expect(command).not.toMatch(/-y$/)
-		expect(command).not.toContain("--agent")
 	})
 })
